+++ src/Oauth2CredentialManagers/CacheStore.php
@@ -1,0 +1,131 @@
+<?php
+
+
+namespace Webfox\Xero\Oauth2CredentialManagers;
+
+
+use Illuminate\Session\Store;
+use Illuminate\Cache\Repository;
+use League\OAuth2\Client\Token\AccessTokenInterface;
+use Webfox\Xero\Oauth2Provider;
+use Webfox\Xero\OauthCredentialManager;
+
+class CacheStore implements OauthCredentialManager
+{
+    /** @var Repository  */
+    protected $cache;
+
+    /** @var Oauth2Provider  */
+    protected $oauthProvider;
+
+    /** @var Store */
+    protected $session;
+
+    protected $cacheKey = 'xero_oauth';
+
+    public function __construct(Repository $cache, Store $session, Oauth2Provider $oauthProvider)
+    {
+        $this->cache         = $cache;
+        $this->oauthProvider = $oauthProvider;
+        $this->session       = $session;
+    }
+
+    public function getAccessToken(): string
+    {
+        return $this->data('token');
+    }
+
+    public function getRefreshToken(): string
+    {
+        return $this->data('refresh_token');
+    }
+
+    public function getTenantId(): string
+    {
+        return $this->data('tenant_id');
+    }
+
+    public function getExpires(): int
+    {
+        return $this->data('expires');
+    }
+
+    public function getState(): string
+    {
+        return $this->session->get($this->cacheKey);
+    }
+
+    public function getAuthorizationUrl(): string
+    {
+        $redirectUrl = $this->oauthProvider->getAuthorizationUrl(['scope' => config('xero.oauth.scopes')]);
+        $this->session->put($this->cacheKey, $this->oauthProvider->getState());
+
+        return $redirectUrl;
+    }
+
+    public function getData(): array
+    {
+        return $this->data();
+    }
+
+    public function exists(): bool
+    {
+        return $this->cache->has($this->cacheKey);
+    }
+
+    public function isExpired(): bool
+    {
+        return time() >= $this->data('expires');
+    }
+
+    public function refresh(): void
+    {
+        $newAccessToken = $this->oauthProvider->getAccessToken('refresh_token', [
+            'refresh_token' => $this->getRefreshToken(),
+        ]);
+
+        $this->store($newAccessToken);
+    }
+
+    public function store(AccessTokenInterface $token, string $tenantId = null): void
+    {
+        $this->cache->forever($this->cacheKey, [
+            'token'         => $token->getToken(),
+            'refresh_token' => $token->getRefreshToken(),
+            'id_token'      => $token->getValues()['id_token'],
+            'expires'       => $token->getExpires(),
+            'tenant_id'     => $tenantId ?? $this->getTenantId()
+        ]);
+    }
+
+    public function getUser(): ?array
+    {
+
+        try {
+            $jwt = new \XeroAPI\XeroPHP\JWTClaims();
+            $jwt->setTokenId($this->data('id_token'));
+            $decodedToken = $jwt->decode();
+
+            return [
+                'given_name'  => $decodedToken->getGivenName(),
+                'family_name' => $decodedToken->getFamilyName(),
+                'email'       => $decodedToken->getEmail(),
+                'user_id'     => $decodedToken->getXeroUserId(),
+                'username'    => $decodedToken->getPreferredUsername(),
+                'session_id'  => $decodedToken->getGlobalSessionId()
+            ];
+        } catch (\Throwable $e) {
+            return null;
+        }
+    }
+
+    protected function data($key = null)
+    {
+        if (!$this->exists()) {
+            throw new \Exception('Xero oauth credentials are missing');
+        }
+
+        $cacheData = $this->cache->get($this->cacheKey);
+        return empty($key) ? $cacheData : ($cacheData[$key] ?? null);
+    }
+}